+++ packages/docs/components/Examples/mention/CustomComponentMentionEditor/CustomComponentMentionEditor.tsx
@@ -45,9 +45,12 @@ export default function CustomComponentMentionEditor(): ReactElement {
   const onOpenChange = useCallback((_open: boolean) => {
     setOpen(_open);
   }, []);
-  const onSearchChange = useCallback(({ value }: { value: string }) => {
-    setSuggestions(defaultSuggestionsFilter(value, mentions));
-  }, []);
+  const onSearchChange = useCallback(
+    ({ trigger, value }: { trigger: string; value: string }) => {
+      setSuggestions(defaultSuggestionsFilter(value, mentions, trigger));
+    },
+    []
+  );
 
   return (
     <div