+++ packages/mention/src/__test__/mentionSuggestionsStrategy.test.ts
@@ -31,22 +31,22 @@ const getBlock = (
 
 describe('test strategy with whitespace support disabled', () => {
   test.each([
-    ['test empty string', '', '@', []],
-    ['trigger only', '@', '@', [[0, 1]]],
-    ['match single word', '@the ', '@', [[0, 4]]],
-    ['should match a word with special characters', '@ęĻŌ', '@', [[0, 4]]],
-    ['should match not match spaces', '@the walking dead', '@', [[0, 4]]],
-    ['match within text', 'a lof @of text', '@', [[5, 9]]],
+    ['test empty string', '', ['@'], []],
+    ['trigger only', '@', ['@'], [[0, 1]]],
+    ['match single word', '@the ', ['@'], [[0, 4]]],
+    ['should match a word with special characters', '@ęĻŌ', ['@'], [[0, 4]]],
+    ['should match not match spaces', '@the walking dead', ['@'], [[0, 4]]],
+    ['match within text', 'a lof @of text', ['@'], [[5, 9]]],
     [
       'should not match if no whitespace before trigger',
       'a lof@of text',
-      '@',
+      ['@'],
       [],
     ],
     [
       'should match multiple mentions',
       '@the @walking @dead',
-      '@',
+      ['@'],
       [
         [0, 4],
         [4, 13],
@@ -69,22 +69,22 @@ describe('test strategy with whitespace support disabled', () => {
 });
 describe('test strategy with whitespace support enabled', () => {
   test.each([
-    ['test empty string', '', '@', []],
-    ['trigger only', '@', '@', [[0, 1]]],
-    ['match single word', '@the', '@', [[0, 4]]],
-    ['match single with following whitespace', '@the ', '@', [[0, 5]]],
-    ['should match a word with special characters', '@ęĻŌ', '@', [[0, 4]]],
-    ['match within text', 'a lof @of text', '@', [[6, 14]]],
+    ['test empty string', '', ['@'], []],
+    ['trigger only', '@', ['@'], [[0, 1]]],
+    ['match single word', '@the', ['@'], [[0, 4]]],
+    ['match single with following whitespace', '@the ', ['@'], [[0, 5]]],
+    ['should match a word with special characters', '@ęĻŌ', ['@'], [[0, 4]]],
+    ['match within text', 'a lof @of text', ['@'], [[6, 14]]],
     [
       'should not match if no whitespace before trigger',
       'a lof@of text',
-      '@',
+      ['@'],
       [],
     ],
     [
       'should match multiple mentions with spaces',
       '@the walking dead tv @the white house',
-      '@',
+      ['@'],
       [
         [0, 21],
         [21, 37],
@@ -93,7 +93,7 @@ describe('test strategy with whitespace support enabled', () => {
     [
       'should match multiple mentions with spaces and special characters',
       '@Thomas Müller @Mario Götze',
-      '@',
+      ['@'],
       [
         [0, 15],
         [15, 27],
@@ -116,8 +116,117 @@ describe('test strategy with whitespace support enabled', () => {
   it('should not match entities', () => {
     const callback = jest.fn();
     const key = genKey();
+    const trigger = ['@'];
     mentionSuggestionsStrategy(
-      '@',
+      trigger,
+      true,
+      defaultRegExp
+    )(
+      getBlock(
+        '@the walking dead tv the white house',
+        [
+          {
+            key: (key as unknown) as number,
+            offset: 20,
+            length: 15,
+          },
+        ],
+        {
+          [key]: {
+            type: 'mention',
+            mutability: 'IMMUTABLE',
+            data: {},
+          },
+        }
+      ),
+      callback
+    );
+    expect(callback).toHaveBeenCalledTimes(1);
+    expect(callback).toHaveBeenLastCalledWith(0, 20);
+  });
+});
+describe('multi mentions test strategy with whitespace support disabled', () => {
+  test.each([
+    ['test empty string', '', ['@', '('], []],
+    ['trigger only', '@', ['@', '('], [[0, 1]]],
+    ['match single word', '@the ( ', ['@', '('], [[0, 4], [4, 6]]],
+    ['should match a word with special characters', '@ęĻŌ ęĻŌ(', ['@', '('], [[0, 4]]],
+    ['should match not match spaces', '@the walking (dead', ['@', '('], [[0, 4], [12, 18]]],
+    ['match within text', 'a (lof @of text', ['@', '('], [[1, 6], [6, 10]]],
+    [
+      'should not match if no whitespace before trigger',
+      'a lof@of text(',
+      ['@', '('],
+      [],
+    ],
+    [
+      'should match multiple mentions',
+      '@the (walking @dead',
+      ['@', '('],
+      [
+        [0, 4],
+        [4, 13],
+        [13, 19],
+      ],
+    ],
+  ])('%s', (_description, text, trigger, result) => {
+    const callback = jest.fn();
+    mentionSuggestionsStrategy(
+      trigger,
+      false,
+      defaultRegExp
+    )(getBlock(text), callback);
+    expect(callback).toHaveBeenCalledTimes(result.length);
+    // eslint-disable-next-line no-plusplus
+    for (let i = 0; i < result.length; ++i) {
+      expect(callback).toHaveBeenNthCalledWith(i + 1, ...result[i]);
+    }
+  });
+});
+describe('multi mentions test strategy with whitespace support enabled', () => {
+  test.each([
+    ['test empty string', '', ['@', '('], []],
+    ['trigger only', '@', ['@', '('], [[0, 1]]],
+    ['match single word', '@the ( ', ['@', '('], [[0, 5], [5, 7]]],
+    ['should match a word with special characters', '@ęĻŌ ęĻŌ(', ['@', '('], [[0, 8]]],
+    ['should match not match spaces', '@the walking (dead', ['@', '('], [[0, 13], [13, 18]]],
+    ['match within text', 'a (lof @of text', ['@', '('], [[2, 7], [7, 15]]],
+    [
+      'should not match if no whitespace before trigger',
+      'a lof@of text(',
+      ['@', '('],
+      [],
+    ],
+    [
+      'should match multiple mentions',
+      '@the (walking @dead',
+      ['@', '('],
+      [
+        [0, 5],
+        [5, 14],
+        [14, 19],
+      ],
+    ],
+  ])('%s', (_description, text, trigger, result) => {
+    const callback = jest.fn();
+    mentionSuggestionsStrategy(
+      trigger,
+      true,
+      defaultRegExp
+    )(getBlock(text), callback);
+    expect(callback).toHaveBeenCalledTimes(result.length);
+    // eslint-disable-next-line no-plusplus
+    for (let i = 0; i < result.length; ++i) {
+      expect(callback).toHaveBeenNthCalledWith(i + 1, ...result[i]);
+    }
+  });
+
+  it('should not match entities', () => {
+    const callback = jest.fn();
+    const key = genKey();
+    const trigger = ['@'];
+    mentionSuggestionsStrategy(
+      trigger,
       true,
       defaultRegExp
     )(