+++ packages/mention/src/utils/defaultSuggestionsFilter.ts
@@ -1,13 +1,18 @@
 // Get the first 5 suggestions that match
 
-import { MentionData } from '..';
+import { MentionData, MultiMentionData } from '..';
 
 const defaultSuggestionsFilter = (
   searchValue: string,
-  suggestions: MentionData[]
+  suggestions: MultiMentionData | MentionData[],
+  trigger?: string
 ): MentionData[] => {
   const value = searchValue.toLowerCase();
-  const filteredSuggestions = suggestions.filter(
+  const triggerSuggestions: MentionData[] =
+    trigger && !Array.isArray(suggestions)
+      ? suggestions[trigger]
+      : suggestions as MentionData[];
+  const filteredSuggestions = triggerSuggestions.filter(
     (suggestion) => !value || suggestion.name.toLowerCase().indexOf(value) > -1
   );
   const length =