+++ CHANGELOG.md
@@ -3,3 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Checkout the CHANGELOG.md of each package …
+
+## to be released
+
+- added support for multiple triggers